The led statemachines and 1 buzzer sound done

+++ project/project1/buzzer.h
@@ -3,7 +3,7 @@
 
 void buzzer_init();
 void buzzer_set_period(short cycles);
-void play_note(int note, int duration);
+void play_note(int note, float duration);
 
 #endif // BUZZER_INCLUDED
